% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_package}
\alias{check_package}
\title{Performes various checks}
\usage{
check_package(path = ".", run_gp = FALSE)
}
\arguments{
\item{path}{the path to the package}

\item{run_gp}{run goodpractice tests}
}
\description{
Performes various checks
}
