% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_data_dtgch_cbem4.R
\docType{data}
\name{df_opt_dtgch_cbem4}
\alias{df_opt_dtgch_cbem4}
\title{Test Dataset Panel Height and Weight, Philippines, scrambled data, not actual.}
\format{
csv
}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S1570677X16300107}
}
\usage{
data(df_opt_dtgch_cbem4)
}
\description{
Generated by *\url{https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_dtgch_cbem4.html}*, post-regression, 4 groups, A, alpha
linear as well as log linear production function estimation results.
}
\examples{
data(df_opt_dtgch_cbem4)
head(df_opt_dtgch_cbem4, 10)
print(as.matrix(df_opt_dtgch_cbem4[,c('alpha_lin', 'alpha_log', 'A_lin', 'A_log', 'beta')]))
}
\references{
“Early life height and weight production functions with endogenous energy and protein inputs”, September 2016, 65-81, Economics & Human Biology, 22, Esteban Puentes, Fan Wang, Jere R. Behrman, Flavio Cunha, John Hoddinott, John A. Maluccio, Linda S. Adair, Judith Borja and Reynaldo Martorell
}
\keyword{datasets}
