% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_dist_gini.R
\name{ff_dist_gini_random_var}
\alias{ff_dist_gini_random_var}
\title{Compute gini coefficient between 0 (equality) and 1 based on a discrete random variable.}
\usage{
ff_dist_gini_random_var(ar_x, ar_prob_of_x, bl_sort_sum = TRUE)
}
\arguments{
\item{ar_x}{array of sclaar values x1, x2, ..., xn-1, xn, sorted.}

\item{ar_prob_of_x}{array of probability mass that sums to one for sequentially each
element of ar_x.}

\item{bl_sort_sum}{boolean if true then finds the unique values of ar_x, sort it, and sum
up the probabilities contained in ar_prob_of_x for each one of the unique values in ar_x}
}
\value{
a scalar value of the gini coefficient
}
\description{
Suppose there are N outcomes from x1, to xn, each with P(x_i) of happening. What is the
gini index/coefficient that summarizes the level of inequality of the population that is
described by this discrete random variable?
}
\examples{
# already sorted example
for (fl_binom_success_prob in seq(0.0001,0.9999,length.out=10)) {
    ar_x <- seq(0, 100, by=1)
    ar_prob_of_x <- dbinom(ar_x, 100, fl_binom_success_prob)
    fl_gini_index <- ff_dist_gini_random_var(ar_x, ar_prob_of_x, bl_sort_sum=FALSE)
    st_print <- paste0('binom p(success)=', fl_binom_success_prob ,
                       ', the fl_gini_index=', fl_gini_index)
    print(st_print)
}
# Example 2a, unsorted
set.seed(123)
ar_x <- c(1,2,3,5,2,1,3)
ar_prob_of_x <- runif(length(ar_x))
ar_prob_of_x <- ar_prob_of_x/sum(ar_prob_of_x)
ff_dist_gini_random_var(ar_x, ar_prob_of_x, bl_sort_sum=TRUE)
# Example 2b, sorted same
ar_x_sort <- c(1,2,3,5)
ar_prob_of_x_sort <- c(0,0,0,0)
ar_prob_of_x_sort[1] <- ar_prob_of_x[1] + ar_prob_of_x[6]
ar_prob_of_x_sort[2] <- ar_prob_of_x[2] + ar_prob_of_x[5]
ar_prob_of_x_sort[3] <- ar_prob_of_x[3] + ar_prob_of_x[7]
ar_prob_of_x_sort[4] <- ar_prob_of_x[4]
ff_dist_gini_random_var(ar_x_sort, ar_prob_of_x_sort, bl_sort_sum=FALSE)
ff_dist_gini_random_var(ar_x_sort, ar_prob_of_x_sort, bl_sort_sum=TRUE)

}
\references{
\url{https://fanwangecon.github.io/REconTools/articles/fv_dist_gini_vector_pos.html}
\url{https://fanwangecon.github.io/R4Econ/math/func_ineq/htmlpdfr/fs_gini_disc.html}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
