% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_dist_gini.R
\name{ff_dist_gini_vector_pos}
\alias{ff_dist_gini_vector_pos}
\title{Compute gini coefficient between 1 (equality) and 2/(n+1) (inequality) where n is vector length.}
\usage{
ff_dist_gini_vector_pos(ar_pos)
}
\arguments{
\item{ar_pos}{array of numeric values that are all non-negative}
}
\value{
a scalar value of the gini coefficient
}
\description{
See vignette for details, but very simple one line algorithm that computes gini for a pos or zero data vector.
This does not work with negative values. Additionally, this is not for a sample, the vector is the population.
Also note while the highest equality is 1, the greatest inequality only asymptotes to 0, but equals to 2/(n+1).
}
\examples{
ar_equal_n2 = c(1,1)
ar_ineql_alittle_n2 = c(1,2)
ar_ineql_somewht_n2 = c(1,2^3)
ar_ineql_alotine_n2 = c(1,2^5)
ar_ineql_veryvry_n2 = c(1,2^8)
ar_ineql_mostmst_n2 = c(1,2^13)
# Beta draw testing
ar_beta_mostrich_n1000 = rbeta(1000, 5, 1)
ar_beta_mostpoor_n1000 = rbeta(1000, 1, 5)
ar_beta_manyrichmanypoor_nomiddle_n1000 = rbeta(1000, 0.5, 0.5)
ff_dist_gini_vector_pos(ar_equal_n2)
ff_dist_gini_vector_pos(ar_ineql_alittle_n2)
ff_dist_gini_vector_pos(ar_ineql_somewht_n2)
ff_dist_gini_vector_pos(ar_ineql_alotine_n2)
ff_dist_gini_vector_pos(ar_ineql_mostmst_n2)
ff_dist_gini_vector_pos(ar_beta_mostrich_n1000)
ff_dist_gini_vector_pos(ar_beta_mostpoor_n1000)
ff_dist_gini_vector_pos(ar_beta_manyrichmanypoor_nomiddle_n1000)
}
\references{
\url{https://fanwangecon.github.io/REconTools/reference/ff_dist_gini_vector_pos.html}
\url{https://fanwangecon.github.io/REconTools/articles/fv_dist_gini_vector_pos.html}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
