% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_dist_integrate.R
\name{ff_dist_integrate_normal}
\alias{ff_dist_integrate_normal}
\title{Trapezoidal rule style discretized normal random variable.}
\usage{
ff_dist_integrate_normal(fl_eps_mean = 0, fl_eps_sd = 1, it_eps = 10)
}
\arguments{
\item{fl_eps_mean}{float mean of the normal distribution}

\item{fl_eps_sd}{float sd of the normal distribution}

\item{it_eps}{integer number of discretized points to return}
}
\value{
a list of normal draw arrays and checks
\itemize{
  \item ar_eps_val - An array of shock draws as Discrete Random Variable
  \item ar_eps_prb - The probability mass for each discrete point
  \item fl_cdf_total_approx - Approximated aggregate probability
  \item fl_mean_approx - Approximated mean given draws array
  \item fl_sd_approx - Approximated standard deviation given draws array
}
}
\description{
Supose epsilon follows the normal distribution. Draw it_eps number of
normal shocks that best approximates the normal distribution.
Outputs a number of statistics to check to quality of approximation
}
\examples{
ff_dist_integrate_normal(fl_eps_mean=60, fl_eps_sd = 20, it_eps=3)
ff_dist_integrate_normal(fl_eps_mean=60, fl_eps_sd = 20, it_eps=5)
ff_dist_integrate_normal(fl_eps_mean=60, fl_eps_sd = 20, it_eps=7)
ff_dist_integrate_normal(fl_eps_mean=60, fl_eps_sd = 20, it_eps=9)
ff_dist_integrate_normal(fl_eps_mean=60, fl_eps_sd = 20, it_eps=11)
ff_dist_integrate_normal(fl_eps_mean=60, fl_eps_sd = 20, it_eps=13)
ff_dist_integrate_normal(fl_eps_mean=60, fl_eps_sd = 20, it_eps=15)
}
\references{
\url{https://fanwangecon.github.io/REconTools/reference/ff_dist_integrate_normal.html}
\url{https://github.com/FanWangEcon/REconTools/blob/master/R/ff_dist_integrate.R}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
