% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_sup_inout.R
\name{ff_sup_clean_rmd_mod}
\alias{ff_sup_clean_rmd_mod}
\title{Modify R4Econ (and others) Rmd files, add Numbered Sections for HTML, and higher Hierarchy}
\usage{
ff_sup_clean_rmd_mod(st_file, it_hierachy_shift = 2, it_toc_depth = 3)
}
\arguments{
\item{st_file}{string texts contents from a R4Econ RMD file.}

\item{it_hierachy_shift}{rmd modification level of promotion from bookdown hierachy
to own file hierarchy}

\item{it_toc_depth}{rmd own file outputs toc levels to show}
}
\value{
modified texts
}
\description{
Modify R4Econ Rmd files, add Numbered Sections for HTML, and higher Hierarchy. Was
not able to, for HTML file, use pandoc arguments to add in TOC. And Need to have two hierarchies
for RMD's individual specific pdf and html files, and the joint bookdown file combining all.
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
