% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_sup_inout.R
\name{ff_sup_git_status_check}
\alias{ff_sup_git_status_check}
\title{Check if local file is different from git version on windows}
\usage{
ff_sup_git_status_check(spt_file)
}
\arguments{
\item{spt_file}{path to file with file name}
}
\value{
a named list of bools and strings
\itemize{
  \item it_status - 1 is new, 2 is modified, 3 no change
  \item st_git_status - string print out of file status
  \item bl_modified - bool 1 if modified
  \item bl_anewfile - bool 1 if new
  \item bl_nochange - bool 1 if no change
}
}
\description{
A file could be new, or cold be modified, identify this by using the git status
command and interpreting the output.
}
\examples{
spt_file = 'C:/Users/fan/Math4Econ/README.md'
ff_sup_git_status_check(spt_file)

}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
