% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_sup_inout.R
\name{ff_sup_lower_hierarchy}
\alias{ff_sup_lower_hierarchy}
\title{Increase the RMD Hierarchy of Existing Files}
\usage{
ff_sup_lower_hierarchy(st_file, it_hierachy_lower_rmd = 1)
}
\arguments{
\item{st_file}{string texts contents from a R4Econ RMD file.}

\item{it_hierachy_lower_rmd}{integer how many pound tiers to promote hierarchy by}
}
\value{
modified texts
}
\description{
Modify Rmd files by increasing pound sign counts. This is needed for the RMD files
converted from mlx, where I can only set file top hierachy to 2 pounds, but should be
three for M4Econ
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
