% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_sup_inout.R
\name{ff_sup_rmd2htmlpdfr}
\alias{ff_sup_rmd2htmlpdfr}
\title{Searchs for Rmd Files in several folders, render to HTML or PDF in another folder}
\usage{
ff_sup_rmd2htmlpdfr(
  ar_spt_root = c("C:/Users/fan/R4Econ/amto/array/",
    "C:/Users/fan/R4Econ/math/integration"),
  ar_spn_skip = c("basics", "integrate"),
  st_rmd_suffix_pattern = ".Rmd",
  st_save_add_suffix = "",
  spt_out_directory = "C:/Users/fan/downloads",
  bl_recursive = TRUE,
  ls_bool_convert = list(bl_pdf = TRUE, bl_html = TRUE, bl_R = TRUE),
  bl_verbose = TRUE
)
}
\arguments{
\item{ar_spt_root}{array of string paths of folders to search in}

\item{ar_spn_skip}{array a string array of names, if path found contains any of the string
in the array, will skip.}

\item{st_rmd_suffix_pattern}{string name of file suffix to search over}

\item{st_save_add_suffix}{string suffix to add to rendered html, pdf files}

\item{spt_out_directory}{string full path of where to store the rendered html, pdf, r files}

\item{bl_recursive}{boolean if to search in folders recursively}

\item{ls_bool_convert}{list of booleans to generate pdf, html and or R file. Generate only HTML for example.
considered included, searched and found}

\item{bl_verbose}{verbose printing more}
}
\value{
a list of string paths of files generated
\itemize{
  \item ls_spt_pdf_generated - a list of pdf file names
  \item ls_spt_html_generated - a list of html file names
  \item ls_spt_R_generated - a list of R file names generated
}#'
}
\description{
searches in several directories for files with Rmd suffix. Skip some file names that contain
strings in ar_spn_skip. Save output rendered HTML and PDF files to a directory. And control
if to render Rmd to PDF, HTML as well as R.
}
\examples{
ar_spt_root <- c('C:/Users/fan/R4Econ/support/rmd')
ar_spn_skip <- c('matrix', 'tibble', '_main')
spt_out_directory <- 'C:/Users/fan/downloads'
st_save_add_suffix <- '_newsave'
ls_bool_convert <- list(bl_pdf=TRUE, bl_html=TRUE, bl_R=TRUE)
bl_verbose <- TRUE
ff_sup_rmd2htmlpdfr(ar_spt_root=ar_spt_root, ar_spn_skip=ar_spn_skip,
                    spt_out_directory=spt_out_directory, st_save_add_suffix=st_save_add_suffix,
                    ls_bool_convert=ls_bool_convert,
                    bl_verbose=bl_verbose)

}
\references{
\url{https://fanwangecon.github.io/R4Econ/development/inout/htmlpdfr/fs_rmd_pdf_html.html}
\url{https://github.com/FanWangEcon/REconTools/blob/master/R/ff_sup_inout.R}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
