% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_requests.R
\name{import_requests}
\alias{import_requests}
\title{Read RECON requests}
\usage{
import_requests(file = "RECON-requests", compile = TRUE, erase = FALSE)
}
\description{
Read RECON requests
}
\author{
Thibaut Jombart
}
