% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_package}
\alias{init_package}
\title{Initializes a new package}
\usage{
init_package(pkg_name, path = ".", check_cran_name = TRUE)
}
\arguments{
\item{pkg_name}{the package name}

\item{path}{the directory path. Default is the current directory}

\item{check_cran_name}{check if the name is already taken on CRAN}
}
\description{
Initializes a new package
}
