% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REconTools-package.R
\docType{package}
\name{REconTools-package}
\alias{REconTools}
\alias{REconTools-package}
\title{REconTools: R Tools for Panel Data and Optimization}
\description{
A R package with reusable functions shared by research projects
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fanwangecon.github.io/REconTools/}
  \item Report bugs at \url{https://github.com/FanWangEcon/REconTools/issues}
}

}
\author{
\strong{Maintainer}: Fan Wang \email{fanwecon@gmail.com}

}
\keyword{internal}
